Ispis greske prilikom ulaza neautentifikovanog korisnika

+++ src/main/java/ftn/project/controller/ClinicController.java
@@ -37,10 +37,24 @@ public class ClinicController {
 	
 	
 	@GetMapping("/clinics")
-	public ModelAndView showClinics(@ModelAttribute("clinicDto")ClinicDto clinicDto,ModelMap model) {
+	public ModelAndView showClinics(HttpServletRequest request,@ModelAttribute("clinicDto")ClinicDto clinicDto,ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return new ModelAndView("badUser");
+		}
+		else {
+			//autorizacija
+			if(userService.getCCA()){
+		
 		model.addAttribute("clinicsDto",clinicService.allClinics());
 	//	model.addAttribute("allClinicsAdmins",userService.allUserByRole("Clinic Administrator"));
 		return new ModelAndView("clinics","Model",clinicService.allClinics());
+			}else {
+				return new ModelAndView("badUser");
+			}
+			}
 	}
 	
 	
@@ -53,24 +67,67 @@ public class ClinicController {
 	
 	@GetMapping("/clinicsProfileCA")
 	public String showClinicProfile(HttpServletRequest request,@ModelAttribute("userDto") UserDto userDto, ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return "redirect:/badUser";
+		}
+		else {
+			//autorizacija
+			if(userService.getCCA()){
+		
 		String username=(String)request.getSession().getAttribute("logUsername");
 		model.addAttribute("clinicDto", clinicService.getClinicProfile(username));
 		return "clinicProfileCA";
+			}else {
+				return "redirect:/badUser";
+			}
+				
+			}
 
 	}
 	
 	@GetMapping("clinics/delete/{idDto}") 
-	public String deleteClinic(@PathVariable("idDto") Long idDto,ModelMap model) {
+	public String deleteClinic(HttpServletRequest request,@PathVariable("idDto") Long idDto,ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return "redirect:badUser";
+		}
+		else {
+			//autorizacija
+			if(userService.getCCA()){
 		clinicService.deleteClinic(idDto);
 		return "redirect:/clinics";
+			}else {
+				return "redirect:badUser";
+			}
+			}
 	}
 	
 	@GetMapping("/clinicProfileCA/edit/{idDto}")
-	public ModelAndView showNurseProfileEdit(@ModelAttribute("clinicDto") ClinicDto clinicDto,@PathVariable("idDto") Long idDto, ModelMap model ) {
+	public ModelAndView showNurseProfileEdit(HttpServletRequest request,@ModelAttribute("clinicDto") ClinicDto clinicDto,@PathVariable("idDto") Long idDto, ModelMap model ) {
+		
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return new ModelAndView("badUser");
+		}
+		else {
+			//autorizacija
+			if(userService.getCCA()){
+		
 		
 		model.addAttribute("clinicDto", clinicService.getClinicById(idDto));
 		return new ModelAndView("clinicProfileCAEdit", "Model",clinicService.getClinicById(idDto));
 
+			}else {
+				return new ModelAndView("badUser");
+			}
+			}
 	}
 	
 	@PostMapping("/clinic/edit/ca")
