+++ src/main/java/ftn/project/controller/RoomController.java
@@ -27,12 +27,28 @@ import lombok.Data;
 @Controller
 public class RoomController  {
 
-	private RoomService roomService;
+	private final RoomService roomService;
+	private final UserService userService;
 
 	@GetMapping("/rooms")
-	public ModelAndView showClinics(@ModelAttribute("roomDto")RoomDto roomDto,ModelMap model) {
+	public ModelAndView showClinics(HttpServletRequest request,@ModelAttribute("roomDto")RoomDto roomDto,ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return new ModelAndView("badUser");
+		}
+		else {
+			//autorizacija
+			if(userService.getCA()){
+
+		
 		model.addAttribute("roomsDto",roomService.allRooms());
 		return new ModelAndView("roomCA","Model",roomService.allRooms());
+			}else {
+				return new ModelAndView("badUser");
+			}
+			}
 	}
 	
 	@PostMapping("/room/create")
@@ -51,14 +67,38 @@ public class RoomController  {
 
 
 	@GetMapping("room/delete/{idDto}")
-	public String deleteRoom(@PathVariable("idDto") Long idDto, ModelMap model) {
+	public String deleteRoom(HttpServletRequest request,@PathVariable("idDto") Long idDto, ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return "redirct;/badUser";
+		}
+		else {
+			//autorizacija
+			if(userService.getCA()){
+		
 		roomService.deleteRoom(idDto);
 		return "redirect:/rooms";
+			}else {
+				return "redirct;/badUser";
+			}
+	}
 	}
 	
 	
 	@GetMapping("/roomsSearch")
 	public ModelAndView searchDoctor(HttpServletRequest request, @ModelAttribute("roomDto") RoomDto roomDto, ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return new ModelAndView("badUser");
+		}
+		else {
+			//autorizacija
+			if(userService.getCA()){
+		
 		ModelAndView mav= new ModelAndView("RoomSearch");
 		
 		Set<RoomDto> sobe=(Set<RoomDto>)request.getSession().getAttribute("roomsDto");
@@ -69,6 +109,10 @@ public class RoomController  {
 			mav.addObject("roomsDto", sobe);
 		}
 		return mav;
+			}else {
+				return new ModelAndView("badUser");
+			}
+			}
 	}
 	
 	@PostMapping("/room/search")
